Remove the renewal lines in terraform output
Terraform output places the changes it will be making between two
lines that look like:

```
---------------------------------------------------------------------

< ... important changes here ...>

---------------------------------------------------------------------
```

Everything outside of these lines is related to renewing modules and
other terraform-specific stuff. These updates are not important for
PRs and the changes will be noted within the `renewal` lines
anyways. I will filter these lines out so the change sets are easier
to read and digest.